# Improve shuffle performance