Using two methods: not using perfectly unbiased ranges (see rationale in the comment), and making better use of the RNG (like not requesting 64 bits on an `usize` when there are less than 2^32 elements).

Before:
```
test misc_shuffle_1000 (x86_64)      ... bench:      10,370 ns/iter (+/- 91)
test misc_shuffle_1000 (x86)         ... bench:       7,565 ns/iter (+/- 45)
```

After:
```
test misc_shuffle_1000 (x86_64)      ... bench:       1,559 ns/iter (+/- 9)
test misc_shuffle_1000 (x86)         ... bench:       2,540 ns/iter (+/- 112)
```

Note that I didn't do endless tries optimizing this, I only implemented the basic ideas. It is a little surprising that x86 does not improve in a way similar to x86_64.