@pitdicker 2018-03-12
Nice to compare against Xorshift.
```
test gen_u32_xorshift      ... bench:       1,374 ns/iter (+/- 6) = 2911 MB/s
```