@dhardy 2018-03-12
Wait, your code can shuffle a 1000-element list in 13% more time than it takes to generate a u32 with Xorshift? I don't believe that.

Also, I stand by my previous comment that algorithms don't belong in the `Rng` trait :wink: Maybe it's time to investigate @vks's idea for an `algorithms` module.

TODO: investigate why the benchmark is so fast. This doesn't need to land in the 0.5 release.