@pitdicker 2018-03-12
I believe the benchmark is right. For a slice with only 1000 elements two values can be swapped with every round of the RNG. Also the only things it does extra are keeping a loop counter, a multiply and the swap itself. If you increase the benchmark to use slices of for example 40mb it takes about twice as long, as expected because it mostly uses the slower loop.