@dhardy 2018-03-12
Am I missing something? `b.iter(...)` should be measuring the time taken to shuffle 1000 elements, which with your code requires 250 `u64` values, which is 500 calls to Xorshift's `next_u32` fn — so the iteration time should be at least 500 times longer than `next_u32` — oh, but `gen_u32_xorshift` calls that 1000 times per iteration; okay.