@pitdicker 2018-03-14
You are getting creative with the labels :smile:.

> I'm not really happy about putting a non-trivial algorithm within `Rng` though. Hopefully later we can find another place for this; in the mean time perhaps a private function would do.

I agree. At the moment I am trying out a single trait that implements functions like https://github.com/dhardy/rand/issues/82#issuecomment-372930308. Curious to see if it works at all...

Also can we ignore this PR for a while? I am getting less convinced by my own arguments that the bias is small :smile:. The range is getting 1 step smaller every time. Is there another way to determine the acceptable `zone` to sample from without division, making use of this pattern?