@dhardy 2018-03-14
Yes, sure, it doesn't really matter if this is left until after 0.5.

Huh, I was trying to make labels actually convey useful information at a glance. Slightly useful maybe.