@pitdicker 2018-03-14
> Huh, I was trying to make labels actually convey useful information at a glance. Slightly useful maybe.

Seems like a success to me