@pitdicker 2018-03-21
I have done several tries with a `ShrinkingRange` distribution. If we can calculate the acceptable zone for the ranges without divisions, would `shuffle` be faster? It is possible with only additions and subtractions after the initial set-up. But the extra instructions and branches turned out to not be really faster, but much more complex.

So I have only kept the idea here to use smaller integers than `usize` if possible, and used the normal `gen_range`.

Benchmark results:
```
test misc_shuffle_1000               ... bench:       4,026 ns/iter (+/- 276)
test misc_shuffle_1000 (x86)         ... bench:       4,630 ns/iter (+/- 42)
```