@pitdicker 2018-03-21
Maybe moving the function to the `seq` module was not the best idea. The entire module is not available without `std`. I would like to make it available, because of the idea's that are forming in https://github.com/dhardy/rand/issues/82, that are mostly `no_std` compatible.

What do you think?