@pitdicker 2018-03-21
Also, if you want to ignore this until after 0.5, feel free to do so.
I just didn't want to leave it a half-finished idea.