@dhardy 2018-03-21
I have some catching up to do, sorry. I'd like to get a 0.5 pre-release cut very soon; I guess non-ABI breaking changes like this could potentially still make the final 0.5 anyway.