@dhardy 2018-03-22
Not sure about the organisation. You could no-std enable seq or just put the private function in `lib.rs` for now; I guess dhardy#82 will result in much of this changing again in the future anyway.