@pitdicker 2018-03-23
A nice property of the changes here: previously `shuffle` was platform-dependent, because it depended on the size of `usize` for the number of bits for every swap. Now it will use 32 bits or less if the slice contains less than 2^31 elements on both 32- and 64-bit platforms. As I believe that is the limit for allocations on 32-bit, that means the `shuffle` operation is now repeatable on all architectures.