@pitdicker 2018-03-31
I am going to close this PR for now. I have kept working in it, and some more of the things from https://github.com/dhardy/rand/issues/82, in https://github.com/pitdicker/rand/commits/slice_random.