@TheIronBorn 2018-06-08
I found another easy performance win for shuffles: unchecked swaps:
```
test misc_shuffle_1000               ... bench:      14,025 ns/iter (+/- 9,005)
test misc_shuffle_unchecked_1000     ... bench:      13,618 ns/iter (+/- 10,550)
```
(I used only `gen_range::<u32>` to reduce overhead)

Perhaps there is some way we could convince the compiler bound checks are not needed. Otherwise this will do:
```rust
fn swap_unchecked<T>(values: &mut [T], a: usize, b: usize) {
    unsafe {
        let pa: *mut T = values.get_unchecked_mut(a);
        let pb: *mut T = values.get_unchecked_mut(b);
        ptr::swap(pa, pb);
    }
}